DX10 Fluid Renderer: code refactoring and improvement
Used range-based for and replaced int with size_t

+++ src/Layers/xrRenderDX10/3DFluid/dx103DFluidManager.h
@@ -84,9 +84,9 @@ private:
     void InitShaders();
     void DestroyShaders();
 
-    void PrepareTexture(int rtIndex);
-    void CreateRTTextureAndViews(int rtIndex, D3D_TEXTURE3D_DESC TexDesc);
-    void DestroyRTTextureAndViews(int rtIndex);
+    void PrepareTexture(size_t rtIndex);
+    void CreateRTTextureAndViews(size_t rtIndex, D3D_TEXTURE3D_DESC TexDesc);
+    void DestroyRTTextureAndViews(size_t rtIndex);
 
     void Reset();
